// src/features/shared/model/enums.ts
export enum Departments {
  HR = 'HR',
  Finance = 'Finance',
  Operations = 'Operations',
  Engineering = 'Engineering',
  Support = 'Support',
  Sales = 'Sales',
  Marketing = 'Marketing',
  Admin = 'Admin',
  Management = 'Management',
}

export enum WorkMode {
  Office = 'office',
  Home = 'home',
}

export enum Remarks {
  EmergencyLeave = 'Emergency Leave',
  AMLeave = 'AM Leave',
  PMLeave = 'PM Leave',
  SpecifyOtherRemark = 'Specify other remark',
}